Mudanças nos textos e dos backgrounds.

+++ lib/presentation/widgets/agendamento_widget.dart
@@ -34,7 +34,7 @@ class AgendamentosWidget extends StatelessWidget {
               trailing: IconButton(
                 icon: const Icon(Icons.delete),
                 onPressed: () {
-                  excluirAgendamento(agendamento.id as int);
+                  excluirAgendamento(agendamento.id);
                 },
               ),
             );
